refactor: layouts 폴더 삭제
